refactor(v2): move core typing to types.ts (#1585)
* refactor(v2): move typing to types.ts

* nits
