+++ packages/docusaurus/src/client/types.ts
@@ -1,0 +1,5 @@
+import {DocusaurusConfig} from '../server/types';
+
+export interface DocusaurusContext {
+  siteConfig?: DocusaurusConfig;
+}
