+++ packages/docusaurus/src/server/loadSetup.ts
@@ -6,7 +6,8 @@
  */
 
 import path from 'path';
-import {load, Props} from './index';
+import {load} from './index';
+import {Props} from './types';
 
 // Helper methods to setup dummy/fake projects
 export const loadSetup = async (name: string): Promise<Props> => {
